requirements.txt
Flask
werkzeug
gevent
tensorflow
keras
numpy
Pillow